+++ src/app/index.ts
@@ -40,15 +40,14 @@ import { ElementSymbolColorThemeParams } from 'Molstar/mol-theme/color/element-s
 import { Overpaint } from 'Molstar/mol-theme/overpaint';
 import { Asset } from 'Molstar/mol-util/assets';
 import { Color } from 'Molstar/mol-util/color/color';
-import { ColorName, ColorNames } from 'Molstar/mol-util/color/names';
 import { RxEventHelper } from 'Molstar/mol-util/rx-event-helper';
 import { CustomEvents } from './custom-events';
 import { PDBeDomainAnnotations } from './domain-annotations/behavior';
 import * as Foldseek from './extensions/foldseek';
-import { AlphafoldView, LigandView, LoadParams, ModelServerRequest, PDBeVolumes, QueryHelper, QueryParam, StructureComponentTags, Tags, addDefaults, applyOverpaint, getStructureUrl, runWithProgressMessage } from './helpers';
+import { AlphafoldView, LigandView, LoadParams, ModelServerRequest, PDBeVolumes, QueryHelper, QueryParam, StructureComponentTags, Tags, addDefaults, applyOverpaint, getStructureUrl, normalizeColor, runWithProgressMessage } from './helpers';
 import { LoadingOverlay } from './overlay';
 import { PluginCustomState } from './plugin-custom-state';
-import { ColorParams, DefaultParams, DefaultPluginUISpec, InitParams, validateInitParams } from './spec';
+import { AnyColor, DefaultParams, DefaultPluginUISpec, InitParams, validateInitParams } from './spec';
 import { initParamsFromHtmlAttributes } from './spec-from-html';
 import { subscribeToComponentEvents } from './subscribe-events';
 import { initSuperposition } from './superposition';
@@ -505,16 +504,7 @@ export class PDBeMolstarPlugin {
     }
 
     normalizeColor(colorVal: AnyColor | null | undefined, defaultColor: Color = Color.fromRgb(170, 170, 170)): Color {
-        try {
-            if (colorVal === undefined || colorVal === null) return defaultColor;
-            if (typeof colorVal === 'number') return Color(colorVal);
-            if (typeof colorVal === 'string' && colorVal[0] === '#') return Color(Number(`0x${colorVal.substring(1)}`));
-            if (typeof colorVal === 'string' && colorVal in ColorNames) return ColorNames[colorVal as ColorName];
-            if (typeof colorVal === 'object') return Color.fromRgb(colorVal.r ?? 0, colorVal.g ?? 0, colorVal.b ?? 0);
-        } catch {
-            // do nothing
-        }
-        return defaultColor;
+        return normalizeColor(colorVal, defaultColor);
     }
 
     /** Get structure ref for a structure with given `structureNumberOrId`.
@@ -567,12 +557,12 @@ export class PDBeMolstarPlugin {
             PluginCommands.Layout.Update(this.plugin, { state: { isExpanded: isExpanded } });
         },
 
-        applySettings: async (settings?: { color?: { r: number, g: number, b: number }, lighting?: string }) => {
+        applySettings: async (settings?: { color?: AnyColor, lighting?: string }) => {
             if (!settings) return;
             if (!this.plugin.canvas3d) return;
             const renderer = { ...this.plugin.canvas3d.props.renderer };
             if (settings.color) {
-                renderer.backgroundColor = Color.fromRgb(settings.color.r, settings.color.g, settings.color.b);
+                renderer.backgroundColor = normalizeColor(settings.color);
             }
             if (settings.lighting) {
                 (renderer as any).style = { name: settings.lighting }; // I don't think this does anything and I don't see how it could ever have worked
@@ -649,7 +639,7 @@ export class PDBeMolstarPlugin {
          * (this will look the same as when the user hovers over a part of the structure).
          * If `focus`, also zoom on the highlighted part.
          * If `structureNumber` is provided, use the specified structure (numbered from 1!); otherwise use the last added structure. */
-        highlight: async (params: { data: QueryParam[], color?: ColorParams, focus?: boolean, structureNumber?: number }) => {
+        highlight: async (params: { data: QueryParam[], color?: AnyColor, focus?: boolean, structureNumber?: number }) => {
             const loci = this.getLociForParams(params.data, params.structureNumber);
             if (Loci.isEmpty(loci)) return;
             if (params.color) {
@@ -672,7 +662,7 @@ export class PDBeMolstarPlugin {
          * If any items in `data` contain `sideChain` or `representation`, add extra representations to them (colored in `representationColor` if provided).
          * If `structureNumber` is provided, apply to the specified structure (numbered from 1!); otherwise apply to all loaded structures.
          * Remove any previously added coloring and extra representations, unless `keepColors` and/or `keepRepresentations` is set. */
-        select: async (params: { data: QueryParam[], nonSelectedColor?: any, structureId?: string, structureNumber?: number, keepColors?: boolean, keepRepresentations?: boolean }) => {
+        select: async (params: { data: QueryParam[], nonSelectedColor?: AnyColor, structureId?: string, structureNumber?: number, keepColors?: boolean, keepRepresentations?: boolean }) => {
             const structureNumberOrId = params.structureId ?? params.structureNumber;
             await this.visual.clearSelection(structureNumberOrId, { keepColors: params.keepColors, keepRepresentations: params.keepRepresentations });
 
@@ -704,7 +694,7 @@ export class PDBeMolstarPlugin {
                     .filter(s => s.param.color !== null)
                     .map(s => ({
                         bundle: s.bundle,
-                        color: s.param.color ? this.normalizeColor(s.param.color) : DefaultSelectColor,
+                        color: s.param.color ? normalizeColor(s.param.color) : DefaultSelectColor,
                         clear: false,
                     }));
                 if (params.nonSelectedColor) {
@@ -712,7 +702,7 @@ export class PDBeMolstarPlugin {
                     if (wholeStructBundle) {
                         overpaintLayers.unshift({
                             bundle: wholeStructBundle,
-                            color: this.normalizeColor(params.nonSelectedColor),
+                            color: normalizeColor(params.nonSelectedColor),
                             clear: false,
                         });
                     }
@@ -725,7 +715,7 @@ export class PDBeMolstarPlugin {
                     if (!bundle) continue;
                     const overpaintLayers: Overpaint.BundleLayer[] = selections.filter(s => s.param.representationColor).map(s => ({
                         bundle: s.bundle,
-                        color: this.normalizeColor(s.param.representationColor!),
+                        color: normalizeColor(s.param.representationColor!),
                         clear: false,
                     }));
                     await this.plugin.build()
@@ -817,19 +807,19 @@ export class PDBeMolstarPlugin {
         /** Set highlight and/or selection color.
          * Highlight color is used when the user hovers over a part of the structure or when applying the `highlight` method.
          * Selection color is used when creating selections with Selection Mode (the mouse cursor icon) and is not related to the color used by the `select` method. */
-        setColor: async (params: { highlight?: ColorParams, select?: ColorParams }) => {
+        setColor: async (params: { highlight?: AnyColor, select?: AnyColor }) => {
             if (!this.plugin.canvas3d) return;
             if (!params.highlight && !params.select) return;
             const renderer = { ...this.plugin.canvas3d.props.renderer };
             const marking = { ...this.plugin.canvas3d.props.marking };
             if (params.highlight) {
-                renderer.highlightColor = this.normalizeColor(params.highlight);
-                marking.highlightEdgeColor = Color.darken(this.normalizeColor(params.highlight), 1);
+                renderer.highlightColor = normalizeColor(params.highlight);
+                marking.highlightEdgeColor = Color.darken(normalizeColor(params.highlight), 1);
                 this.isHighlightColorUpdated = true;
             }
             if (params.select) {
-                renderer.selectColor = this.normalizeColor(params.select);
-                marking.selectEdgeColor = Color.darken(this.normalizeColor(params.select), 1);
+                renderer.selectColor = normalizeColor(params.select);
+                marking.selectEdgeColor = Color.darken(normalizeColor(params.select), 1);
                 this.isSelectedColorUpdated = true;
             }
             await PluginCommands.Canvas3D.SetSettings(this.plugin, { settings: { renderer, marking } });
@@ -928,9 +918,6 @@ export class PDBeMolstarPlugin {
         foldseek: Foldseek,
     };
 }
-
-
-type AnyColor = ColorParams | string | number
 
 
 (window as any).PDBeMolstarPlugin = PDBeMolstarPlugin;